requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.18.0
langchain>=0.2.0
langchain-core>=0.2.0
langgraph>=0.2.0
openai>=1.50.0
python-dotenv>=1.0.0
bcrypt>=4.1.0
openpyxl>=3.1.0
reportlab>=4.1.0
Pillow>=10.2.0
kaleido>=0.2.1
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymysql>=1.1.0 